requirements.txt
openai
langchain 
langchain_community
streamlit 
python-dotenv
PyPDF2

-e .